app-mgmt renamed to app-info
