+++ charts/meep-app-enablement/values-template.yaml
@@ -30,14 +30,14 @@ ingress:
       paths:
         - /{{ .SandboxName }}/mec_app_support
         - /{{ .SandboxName }}/mec_service_mgmt
-        - /{{ .SandboxName }}/mec_app_mgmt
+        - /{{ .SandboxName }}/app_info
   annotations:
     kubernetes.io/ingress.class: nginx
     nginx.ingress.kubernetes.io/force-ssl-redirect: {{ .HttpsOnly }}
     nginx.ingress.kubernetes.io/configuration-snippet: |
       rewrite ^/{{ .SandboxName }}/mec_app_support(/|$)(.*)$ /mec_app_support/$2 break;
       rewrite ^/{{ .SandboxName }}/mec_service_mgmt(/|$)(.*)$ /mec_service_mgmt/$2 break;
-      rewrite ^/{{ .SandboxName }}/mec_app_mgmt(/|$)(.*)$ /mec_app_mgmt/$2 break;
+      rewrite ^/{{ .SandboxName }}/app_info(/|$)(.*)$ /app_info/$2 break;
     {{- if .AuthEnabled }}
     nginx.ingress.kubernetes.io/auth-url: https://$http_host/auth/v1/authenticate?svc=meep-app-enablement&sbox={{ .SandboxName }}
     {{- end }}
