+++ go-apps/meep-app-enablement/server/app-info/README.md
@@ -1,6 +1,6 @@
 # Go API Server for server
 
-AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
 
 ## Overview
 This server was generated by the [swagger-codegen]
@@ -13,7 +13,7 @@ To see how to make this your own, look here:
 [README](https://github.com/swagger-api/swagger-codegen/blob/master/README.md)
 
 - API version: 1.0.0
-- Build date: 2021-06-18T14:51:08.832-04:00[America/New_York]
+- Build date: 2021-06-21T14:08:01.779-04:00[America/New_York]
 
 
 ### Running the server
