+++ go-apps/meep-app-enablement/server/app-info/app-info.go
