+++ go-apps/meep-app-enablement/server/app-info/convert.go
