+++ go-apps/meep-app-enablement/server/app-info/model_application_info.go
@@ -13,21 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * AdvantEDGE MEC Application Management API
+ * AdvantEDGE Application Information API
  *
- * AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+ * AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
  *
  * API version: 1.0.0
- * Contact: cti_support@etsi.org
+ * Contact: AdvantEDGE@InterDigital.com
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 package server
 
-// ApplicationState : This enumeration defines the possible states of an application.
-type ApplicationState string
+// This type represents the general information of a MEC application.
+type ApplicationInfo struct {
+	// Application Instance Id
+	AppInstanceId string `json:"appInstanceId"`
+	// Application Name
+	AppName string `json:"appName,omitempty"`
+	// Application Version
+	Version string `json:"version,omitempty"`
 
-// List of ApplicationState
-const (
-	APPLICATION_STATE_ACTIVE   ApplicationState = "ACTIVE"
-	APPLICATION_STATE_INACTIVE ApplicationState = "INACTIVE"
-)
+	State *ApplicationState `json:"state"`
+}
