+++ go-apps/meep-app-enablement/server/app-info/model_locality_type.go
@@ -13,22 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * AdvantEDGE MEC Application Management API
+ * AdvantEDGE Application Information API
  *
- * AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+ * AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
  *
  * API version: 1.0.0
- * Contact: cti_support@etsi.org
+ * Contact: AdvantEDGE@InterDigital.com
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
+package server
 
-package client
+// LocalityType : The scope of locality as expressed by \"consumedLocalOnly\" and \"isLocal\". If absent, defaults to MEC_HOST
+type LocalityType string
 
-// ApplicationState : This enumeration defines the possible states of an application.
-type ApplicationState string
-
-// List of ApplicationState
+// List of LocalityType
 const (
-	ACTIVE_ApplicationState   ApplicationState = "ACTIVE"
-	INACTIVE_ApplicationState ApplicationState = "INACTIVE"
+	MEC_SYSTEM LocalityType = "MEC_SYSTEM"
+	MEC_HOST   LocalityType = "MEC_HOST"
+	NFVI_POP   LocalityType = "NFVI_POP"
+	ZONE       LocalityType = "ZONE"
+	ZONE_GROUP LocalityType = "ZONE_GROUP"
+	NFVI_NODE  LocalityType = "NFVI_NODE"
 )
