+++ go-apps/meep-app-enablement/server/app-info/model_problem_details.go
@@ -13,24 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * AdvantEDGE MEC Application Management API
+ * AdvantEDGE Application Information API
  *
- * AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+ * AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
  *
  * API version: 1.0.0
- * Contact: cti_support@etsi.org
+ * Contact: AdvantEDGE@InterDigital.com
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
+package server
 
-package client
+type ProblemDetails struct {
+	Type_ string `json:"type,omitempty"`
 
-// This type represents the general information of a MEC application.
-type ApplicationInfo struct {
-	// Application Instance Id
-	AppInstanceId string `json:"appInstanceId"`
-	// Application Name
-	AppName string `json:"appName,omitempty"`
-	// Application Version
-	Version string            `json:"version,omitempty"`
-	State   *ApplicationState `json:"state"`
+	Title string `json:"title,omitempty"`
+
+	Status int32 `json:"status,omitempty"`
+
+	Detail string `json:"detail,omitempty"`
+
+	Instance string `json:"instance,omitempty"`
 }
