+++ go-packages/meep-app-info-client/README.md
@@ -1,6 +1,6 @@
 # Go API client for client
 
-AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
 
 ## Overview
 This API client was generated by the [swagger-codegen](https://github.com/swagger-api/swagger-codegen) project.  By using the [swagger-spec](https://github.com/swagger-api/swagger-spec) from a remote server, you can easily generate an API client.
@@ -17,7 +17,7 @@ import "./client"
 
 ## Documentation for API Endpoints
 
-All URIs are relative to *https://localhost/sandboxname/mec_app_mgmt/v1*
+All URIs are relative to *https://localhost/sandboxname/app_info/v1*
 
 Class | Method | HTTP request | Description
 ------------ | ------------- | ------------- | -------------
@@ -42,5 +42,5 @@ Class | Method | HTTP request | Description
 
 ## Author
 
-cti_support@etsi.org
+AdvantEDGE@InterDigital.com
