+++ go-packages/meep-app-info-client/docs/ApplicationInfo.md
