+++ go-packages/meep-app-info-client/docs/ApplicationState.md
