+++ go-packages/meep-app-info-client/docs/LocalityType.md
