+++ go-packages/meep-app-info-client/model_locality_type.go
@@ -13,24 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * AdvantEDGE MEC Application Management API
+ * AdvantEDGE Application Information API
  *
- * AdvantEDGE implementation to create a Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
+ * AdvantEDGE implementation to create an Application Instance information using OpenAPI. Developed as an extension to Application Enablement API.
  *
  * API version: 1.0.0
- * Contact: cti_support@etsi.org
+ * Contact: AdvantEDGE@InterDigital.com
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
-package server
 
-type ProblemDetails struct {
-	Type_ string `json:"type,omitempty"`
+package client
 
-	Title string `json:"title,omitempty"`
+// LocalityType : The scope of locality as expressed by \"consumedLocalOnly\" and \"isLocal\". If absent, defaults to MEC_HOST
+type LocalityType string
 
-	Status int32 `json:"status,omitempty"`
-
-	Detail string `json:"detail,omitempty"`
-
-	Instance string `json:"instance,omitempty"`
-}
+// List of LocalityType
+const (
+	MEC_SYSTEM_LocalityType LocalityType = "MEC_SYSTEM"
+	MEC_HOST_LocalityType   LocalityType = "MEC_HOST"
+	NFVI_POP_LocalityType   LocalityType = "NFVI_POP"
+	ZONE_LocalityType       LocalityType = "ZONE"
+	ZONE_GROUP_LocalityType LocalityType = "ZONE_GROUP"
+	NFVI_NODE_LocalityType  LocalityType = "NFVI_NODE"
+)
